# openapi.yaml
openapi: 3.0.0
info:
  title: Eterna
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8000
...
paths:
...
  /api/categories/:
    post:
      tags:
        - Categories
      summary: Categories
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                name: JS
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{token}}
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
        - name: Accept
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...
  /api/categories:
    get:
      tags:
        - Categories
      summary: Categories List
      parameters:
        - name: Authorization
          in: header
          schema:
            type: string
          example: Bearer {{token}}
        - name: Accept
          in: header
          schema:
            type: string
          example: application/json
        - name: Content-Type
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...